Create plan model and refactor float type to int

+++ src-tauri/src/models/inventory.rs
@@ -4,7 +4,7 @@ use sqlx::prelude::FromRow;
 #[derive(Debug, Deserialize, FromRow, Serialize, Clone)]
 pub struct Inventory {
     pub id: String,
-    pub amount: f64,
+    pub amount: i64,
     pub item_id: String,
 }
 
@@ -14,7 +14,7 @@ pub struct InventoryWithItem {
     pub item_name: String,
     pub item_picture: Option<String>,
     pub item_id: String,
-    pub amount: Option<f64>,
+    pub amount: Option<i64>,
     pub game_id: String
 }
 
@@ -22,6 +22,6 @@ pub struct InventoryWithItem {
 #[derive(Debug, Deserialize, FromRow, Serialize, Clone)]
 pub struct InventoryDto {
     pub id: Option<String>,
-    pub amount: f64,
+    pub amount: i64,
     pub item_id: String,
 }
